Rename GUI script and update references
